Interim commit with refactor for other sitemap methods
